bolttrip/requirements.txt
Django==4.2.1
psycopg2-binary==2.9.9
djangorestframework
djangorestframework-simplejwt
drf-spectacular
celery
redis
python-dotenv
django-cors-headers
gunicorn
whitenoise
